+++ nodejs/hello/src/lib.rs
@@ -1,3 +1,7 @@
+mod error;
+
+use error::Error;
+
 #[macro_use]
 extern crate napi_derive;
 
@@ -5,19 +9,6 @@ use napi::{
     CallContext, ContextlessResult, Env, JsNumber, JsObject, JsString, JsUndefined, Result,
 };
 use std::convert::TryInto;
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum Error {
-    #[error(transparent)]
-    Hello(#[from] hello::Error),
-}
-
-impl From<Error> for napi::Error {
-    fn from(err: Error) -> Self {
-        Self::from_reason(err.to_string())
-    }
-}
 
 #[module_exports]
 fn init(mut exports: JsObject) -> Result<()> {
@@ -67,7 +58,7 @@ fn concat(ctx: CallContext) -> Result<JsObject> {
 }
 
 #[contextless_function]
-fn raise_error(env: Env) -> ContextlessResult<JsUndefined> {
+pub fn raise_error(env: Env) -> ContextlessResult<JsUndefined> {
     hello::raise_error().map_err(Into::<Error>::into)?;
     env.get_undefined().map(Some)
 }
